Refactored notes to component and  to change link according to mapper
