+++ src/editor-demo/common/doc-texts/installation.js
